% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplotOr.R
\name{barplotOr}
\alias{barplotOr}
\title{Ordered barplots}
\usage{
barplotOr(height, col, decr = TRUE, names.arg = "", ...)
}
\arguments{
\item{height}{see ?barplot}

\item{col}{The colors of the bars. This argument must be provided.}

\item{decr}{A boolean indicating the decreasing (TRUE) or increasing (FALSE)
order.}

\item{names.arg}{see ?barplot}

\item{...}{Additional arguments, see ?barplot.}
}
\value{
A vector with bar positions like in the original barplot function.
}
\description{
Ordered barplots
}
\details{
Function to make a barplot from a matrix with argument
  beside = FALSE with each column ordered according to its value.
}
\examples{
set.seed(1234)
dataset <- matrix(sample(1:20, 104, replace = TRUE), ncol = 26)
myCol <- c("#1B9E77", "#D95F02", "#7570B3", "#E7298A")
graphics::barplot(height = dataset, col = myCol, names.arg = LETTERS)
barplotOr(height = dataset, col = myCol, names.arg = LETTERS)
}
